started restructure of fe-problem.h
